add test on calculateWinner function
